adeo.ts - fz.scenes_recall_scene_6502, fz.adeo_button_65024 and fz.color_stop_raw
